Add a pre-push command, with a push alias for backwards compat
